Implementatio game's logic
First cheking is the input value is empty, its to avoid make unecessary comparisons. If its empty, then display a message to user to enter a number.

Uising if statment to check for winner conditions
Check 3 conditions one if its equal to the generated number, if true display a winner message and change background color

If the user number is biger than generated number it display a message inform that the number is to high
 the last condtions do the opposite